Moving ComponentMap, ComponentSet to pyomo.common.collections
