+++ pyomo/common/collections/__init__.py
@@ -9,3 +9,7 @@
 #  ___________________________________________________________________________
 
 from .orderedset import OrderedDict, OrderedSet
+from .hashabletuple import HashableTuple
+from .component_map import ComponentMap
+from .component_set import ComponentSet
+
